requirements.txt
minio
psycopg2-binary
pymongo
trino
pyspark==3.3.1
